+++ src/main/java/com/ktds/dsquare/board/card/dto/CardResponse.java
@@ -28,12 +28,12 @@ public class CardResponse {
     private LocalDateTime createDate;
     private LocalDateTime lastUpdateDate;
     private Long viewCnt;
-    private Integer likeCnt;
+    private Long likeCnt;
     private Boolean likeYn;
     private CardSelectionInfo selectionInfo;
     private Long commentCnt;
 
-    public static CardResponse toDto(Card entity, Member writer, Team team, Member cardOwner, Integer likeCnt, Boolean likeYn, Long commentCnt) {
+    public static CardResponse toDto(Card entity, Member writer, Team team, Member cardOwner, Long likeCnt, Boolean likeYn, Long commentCnt) {
         MemberInfo writerInfo = MemberInfo.toDto(writer);
         TeamInfo teamInfo = TeamInfo.toDto(team);
         CardSelectionInfo selectionInfo;
